ui: Fix color of statusbar in loading screen.
In android the color of status-bar doesn't match with the background
color in loading screen. This is fixed by adding ZulipStatusBar
component with the props
backgroundColor={Color(BRAND_COLOR).lighten(0.1)}

lighten(0.1) was added because the ZulipStatusBar automatically
darkens any given background-color by 0.1, therefore passing a
color after lightening it by 0.1 balances this effect and gives the
desired output.

Fixes: #3860.